requirements.txt
langchain
langgraph
autogen
requests
python-dotenv
gradio
streamlit
pandas
pytest
pytest-cov
tenacity
aiohttp
huggingface_hub
groq
graphviz
numpy
langchain-core
langchain-community
langsmith>=0.1.75
rich==13.7.1




